+++ NEWS.md
@@ -3,6 +3,10 @@
 
 * Aesthetics are now documented in the manual pages (#28).
 
+* `geom_errorbarh()` now supports `height` for compatibility with
+  ggplot2's version. It no longer matters whether ggplot2 or ggstance
+  is loaded first (#27, #29).
+
 
 # ggstance 0.3.1